requirements.txt
requests
python-dotenv
rich
streamlit